requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
botocore==1.34.86
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.110.1
frozenlist==1.4.1
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
multidict==6.0.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-ldap @ file:///C:/Users/fabio_mattes/Downloads/python_ldap-3.4.4-cp312-cp312-win_amd64.whl#sha256=94d2ca2b3ced81c9d89aa5c79d4965d03053e1ffdcfae73e9fac85d25b692e85
python-multipart==0.0.9
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
websockets==11.0.3
yarl==1.9.4
